+++ KokiWeapons/Components/TPTrap/TPTrap.cs
@@ -35,15 +35,21 @@ public class TPTrap : MonoBehaviour
         || !InstanceFinder.IsServer) return;
 
         detonated = true;
-        var playerIDs = GetPlayersToTeleport();
+        var players = GetPlayers();
         otherTrap.TryGetComponent(out TPTrap otherTrapComponent);
         if (!otherTrapComponent.detonated)
         {
             otherTrapComponent.detonated = true;
-            var otherPlayerIDs = otherTrapComponent.GetPlayersToTeleport();
+            var otherPlayers = otherTrapComponent.GetPlayers();
 
-            playerIDs.Do(ID => SendRPCIfNotHost(ID, otherTrap.transform.position));
-            otherPlayerIDs.Do(ID => SendRPCIfNotHost(ID, this.transform.position));
+            // this is SO poorly written
+            var playerIDs = players.Where(player => TeleportIfHost(player, otherTrap.transform.position))
+            .Select(go => GOToCSteamID(go));
+            var otherPlayerIDs = otherPlayers.Where(player => TeleportIfHost(player, this.transform.position))
+            .Select(go => GOToCSteamID(go));
+
+            playerIDs.Do(ID => TPTrapNetworking.TargetedRPC(ID, "TPPlayer", [otherTrap.transform.position]));
+            otherPlayerIDs.Do(ID => TPTrapNetworking.TargetedRPC(ID, "TPPlayer", [this.transform.position]));
 
             var nob1 = this.gameObject.GetComponent<NetworkObject>().ObjectId;
             var nob2 = otherTrap.GetComponent<NetworkObject>().ObjectId;
@@ -55,23 +61,31 @@ public class TPTrap : MonoBehaviour
         }
     }
 
-    /// The host can't targetRPC themselves.
-    private void SendRPCIfNotHost(CSteamID ID, Vector3 destination)
+    private CSteamID GOToCSteamID(GameObject go)
     {
-        if (ID == SteamUser.GetSteamID())
-            TPTrapNetworking.TPPlayer(destination);
-        else
-            TPTrapNetworking.TargetedRPC(ID, "TPPlayer", [destination]);
+        return (CSteamID)ulong.Parse(go
+                .GetComponent<NetworkObject>().Owner.GetAddress());
     }
 
-    private CSteamID[] GetPlayersToTeleport()
+    private bool TeleportIfHost(GameObject potentialHost, Vector3 pos)
+    {
+        var ID = GOToCSteamID(potentialHost);
+        if (ID == SteamUser.GetSteamID())
+        {
+            potentialHost.GetComponent<FirstPersonController>().Teleport(pos, angle: 0f, boost: false, cac: null, power: 0, decel: 0, dontTranslateRotation: true);
+            return false;
+        }
+        return true;
+    }
+
+    private GameObject[] GetPlayers()
     {
         var boxExtents = new Vector3(3.2f, 1.631936f, 3.2f);
         var layerMask = 1 << 11 | 1 << 16;
-        return Physics.OverlapBox(this.transform.position, boxExtents, Quaternion.identity, layerMask)
+        var colliders = Physics.OverlapBox(this.transform.position, boxExtents, Quaternion.identity, layerMask);
+        return colliders
             .Select(col => col.transform.root.gameObject)
             .Where(go => go.CompareTag("Player"))
-            .Select(go => (CSteamID)ulong.Parse(go.GetComponent<NetworkObject>().Owner.GetAddress()))
             .Distinct()
             .ToArray();
     }